requirements.txt
py-cord
python-dotenv
chat_exporter
youtube-dl
PyNaCl
pymysql
wavelink